+++ docs/CHANGE_LOG.md
@@ -8,6 +8,9 @@
 #### **View: Statistics:**
 - The bars shown in the view now show a tooltip, which shows the total count for each range type.
 
+#### **Public Functions:**
+- Added new public function "switchView()", which will switch the current view for any rendered element.
+
 #### **Fixes:**
 - Fixed the years-list having extra padding around the left, right, and bottom, which caused the menu to stay open longer when the mouse leaves the menu.
 - Fixed the "Chart" view lines being positioned using "margin-left" instead of just "left".